requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
astroid==2.11.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.1.7
bleach==4.1.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
cryptography==2.8
cssselect==1.1.0
decorator==5.1.1
dicttoxml==1.7.4
distro==1.4.0
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-heroku==0.3.1
django-jazzmin==2.5.0
django-js-asset==1.2.2
django-markdown==0.8.4
django-taggit==2.1.0
docutils==0.18.1
entrypoints==0.3
fasteners==0.14.1
frozenlist==1.3.0
future==0.18.2
geocoder
geoip2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.14.0
idna
importlib-metadata==4.10.1
IP2Location==8.7.2
isort==5.10.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==1.1.0
maxminddb==2.2.0
mccabe==0.7.0
monotonic==1.5
multidict==6.0.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pip-review==1.1.1
pkginfo==1.8.2
platformdirs==2.5.1
protobuf==3.6.1
psycopg2==2.8.4
pycairo==1.16.2
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==1.7.1
pylint
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyquery==1.4.3
pyRFC3339==1.1
python-dateutil==2.7.3
python-decouple==3.6
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
ratelim==0.1.6
readme-renderer==34.0
reportlab==3.5.34
requests
requests-toolbelt==0.9.1
requests-unixsocket
rfc3986==2.0.0
SecretStorage==2.3.1
selenium==4.1.3
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.63.0
trio
trio-websocket==0.9.2
twine
typed-ast==1.5.2
typing==3.7.4.3
typing-extensions==4.1.1
urllib3
wadllib==1.3.3
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.14.0
wsproto==1.1.0
yarl
zipp==3.7.0